Rearranged the endpoints to be in separate folders
