--- api/ExpressedRealms.Expressions.API/ExpressionEndpoints/Responses/ExpressionHeaderResponse.cs
@@ -1,7 +1,0 @@
-namespace ExpressedRealms.Expressions.API.ExpressionEndpoints.Responses;
-
-public class ExpressionHeaderResponse
-{
-    public string Name { get; set; } = null!;
-    public string Content { get; set; } = null!;
-}
