+++ api/ExpressedRealms.Expressions.API/Helpers/ExpressionHelpers.cs
@@ -1,8 +1,8 @@
-using ExpressedRealms.Expressions.API.ExpressionEndpoints.DTOs;
 using ExpressedRealms.Expressions.API.ExpressionEndpoints.Requests;
+using ExpressedRealms.Expressions.API.ExpressionSubSectionEndpoints.DTOs;
 using ExpressedRealms.Expressions.Repository.ExpressionTextSections.DTOs;
 
-namespace ExpressedRealms.Expressions.API.ExpressionEndpoints.Helpers;
+namespace ExpressedRealms.Expressions.API.Helpers;
 
 internal static class ExpressionHelpers
 {
